// src/router/router.types.ts
export enum ERoutesPaths {
  HOME = '/',
  COMPANIES = '/companies',
  REGISTER = '/register',
  LOGIN = '/login',
  ABOUT = '/about',
  NOT_FOUND = '/:pathMatch(.*)*',
}

export enum ERoutesNames {
  HOME = 'Home',
  COMPANIES = 'Companies',
  REGISTER = 'Register',
  LOGIN = 'Login',
  ABOUT = 'About',
  NOT_FOUND = 'Not-found',
}